pyproject.toml
[project]
name = "austin-animal-center-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-bigquery>=1.11.1",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "pandas>=3.0.2",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
